src/requirements.txt
keras==2.2.4
tensorflow==1.13.1
nose==1.3.*
pillow==5.0.*
Flask==1.0.*
